Added a mini-report.  Close to first release
All logic should be gone over. SKSPD seems to be buggy along with
interactions between instants and weaponskills.  For example, the time
between the first weaponskill and second weaponskill is usually around
~3.5 seconds but it should be <2.5 seconds.  Actual DPS is scary close
to a real  number. Getting ~220 DPS with my stats, not including
skillspeed which seems to be bugged.

To see the report, go to the bottom of the log.  Should probably go over
and add comments to everything for future reference :X